requirements.txt
pandas
seaborn
protobuf
matplotlib
python-dotenv
